+++ include/rcl_like_wrapper.hpp
@@ -48,16 +48,16 @@ public:
 using MessageTypes = std::unordered_map<std::string, MessageType>;
 
 
-int64_t node_create_node(int64_t domain_id);
-void node_destroy_node(int64_t node_ptr);
-void node_spin(int64_t node_ptr);
-void node_spin_once(int64_t node_ptr);
-void node_stop_spin(int64_t node_ptr);
-int64_t publisher_create_publisher(int64_t node_ptr, std::string message_type_name, std::string topic, eprosima::fastdds::dds::TopicQos& qos);
-void publisher_publish_impl(int64_t publisher_ptr, void* message);
-int32_t publisher_get_subscription_count(int64_t publisher_ptr);
-int64_t subscription_create_subscription(int64_t node_ptr, std::string message_type_name, std::string topic, eprosima::fastdds::dds::TopicQos& qos, std::function<void(void*)> callback);
-int32_t subscription_get_publisher_count(int64_t subscription_ptr);
+intptr_t node_create_node(uint16_t domain_id);
+void node_destroy_node(intptr_t node_ptr);
+void node_spin(intptr_t node_ptr);
+void node_spin_once(intptr_t node_ptr);
+void node_stop_spin(intptr_t node_ptr);
+intptr_t publisher_create_publisher(intptr_t node_ptr, std::string message_type_name, std::string topic, eprosima::fastdds::dds::TopicQos& qos);
+void publisher_publish_impl(intptr_t publisher_ptr, void* message);
+int32_t publisher_get_subscription_count(intptr_t publisher_ptr);
+intptr_t subscription_create_subscription(intptr_t node_ptr, std::string message_type_name, std::string topic, eprosima::fastdds::dds::TopicQos& qos, std::function<void(void*)> callback);
+int32_t subscription_get_publisher_count(intptr_t subscription_ptr);
 void rcl_like_wrapper_init(const MessageTypes &types);
 
 } // namespace rcl_like_wrapper
