+++ src/rcl_like_wrapper.cpp
@@ -9,52 +9,52 @@ namespace {
 }
 
 // Node
-int64_t node_create_node(int64_t domain_id) {
-  return reinterpret_cast<int64_t>(new Node(domain_id));
+intptr_t node_create_node(uint16_t domain_id) {
+  return reinterpret_cast<intptr_t>(new Node(domain_id));
 }
 
-void node_destroy_node(int64_t node_ptr) {
+void node_destroy_node(intptr_t node_ptr) {
   auto node = reinterpret_cast<Node *>(node_ptr);
   node->destroy();
 }
 
-void node_spin(int64_t node_ptr) {
+void node_spin(intptr_t node_ptr) {
   auto node = reinterpret_cast<Node *>(node_ptr);
   node->spin();
 }
 
-void node_spin_once(int64_t node_ptr) {
+void node_spin_once(intptr_t node_ptr) {
   auto node = reinterpret_cast<Node *>(node_ptr);
   node->spin_once();
 }
 
-void node_stop_spin(int64_t node_ptr) {
+void node_stop_spin(intptr_t node_ptr) {
   auto node = reinterpret_cast<Node *>(node_ptr);
   node->stop_spin();
 }
 
 // Publisher
-int64_t publisher_create_publisher(int64_t node_ptr, std::string message_type_name, std::string topic,
+intptr_t publisher_create_publisher(intptr_t node_ptr, std::string message_type_name, std::string topic,
                                  dds::TopicQos& qos) {
   auto node = reinterpret_cast<Node *>(node_ptr);
   // Creating a publisher with the specified message type, topic, and QoS.
-  return reinterpret_cast<int64_t>(node->create_publisher(message_types.at(message_type_name), std::string("rt/") + topic, qos));
+  return reinterpret_cast<intptr_t>(node->create_publisher(message_types.at(message_type_name), std::string("rt/") + topic, qos));
 }
 
-void publisher_publish_impl(int64_t publisher_ptr, void* message) {
+void publisher_publish_impl(intptr_t publisher_ptr, void* message) {
   auto publisher = reinterpret_cast<Publisher *>(publisher_ptr);
   // Publishing the message using the specified publisher.
   publisher->publish(message);
 }
 
-int32_t publisher_get_subscription_count(int64_t publisher_ptr) {
+int32_t publisher_get_subscription_count(intptr_t publisher_ptr) {
   auto publisher = reinterpret_cast<Publisher *>(publisher_ptr);
   // Getting the subscription count associated with the publisher.
   return publisher->get_subscription_count();
 }
 
 // Subscription
-int64_t subscription_create_subscription(int64_t node_ptr, std::string message_type_name, std::string topic, dds::TopicQos& qos, std::function<void(void*)> callback) {
+intptr_t subscription_create_subscription(intptr_t node_ptr, std::string message_type_name, std::string topic, dds::TopicQos& qos, std::function<void(void*)> callback) {
     auto node = reinterpret_cast<Node*>(node_ptr);
     MessageType& message_type = message_types.at(message_type_name);
 
@@ -64,10 +64,10 @@ int64_t subscription_create_subscription(int64_t node_ptr, std::string message_t
         callback(message_data);
     });
 
-    return reinterpret_cast<int64_t>(subscription);
+    return reinterpret_cast<intptr_t>(subscription);
 }
 
-int32_t subscription_get_publisher_count(int64_t subscription_ptr) {
+int32_t subscription_get_publisher_count(intptr_t subscription_ptr) {
   auto subscription = reinterpret_cast<Subscription *>(subscription_ptr);
   // Getting the publisher count associated with the subscription.
   return subscription->get_publisher_count();
@@ -78,5 +78,4 @@ void rcl_like_wrapper_init(const MessageTypes &types) {
   message_types = types;
 }
 
-}
-// namespace rcl_like_wrapper
+}  // namespace rcl_like_wrapper
